+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 - The function for determining the client's IP address has been corrected and improved.
+- ESI: Fixed handling of "character not found" errors (they changed a bit).
 - The `install.sh` and `dist.sh` scripts have been removed; only the Docker development environment
   is now supported.